Step 8: House the config views in ui/ boundaries instead of fencing around them
SetupSteps -> backends/shared/ui/, ClaudeConfigView -> backends/claude/ui/,
CodexConfigView -> backends/codex/ui/, dropping the gallery-fence regex
exemptions. Extract useBackendAuthState into session/ — it is written
purely against the descriptor.auth contract and is consumed from both ui/
and a backend's Configure dialog, which also retires the inline
boundaries/dependencies disable in ClaudeInstallModal.
