+++ src/agentMode/backends/shared/ui/SetupSteps.test.tsx
