+++ src/agentMode/backends/shared/ui/SetupSteps.tsx
